+++ CHANGELOG.md
@@ -44,6 +44,7 @@ All notable changes to Sourcegraph Server and Data Center are documented in this
 * The new `auth.providers` site config is an array of authentication provider objects. Currently only 1 auth provider is supported. The singular `auth.provider` is deprecated.
 * Users authenticated with OpenID Connect are now able to sign out of Sourcegraph (if the provider supports token revocation or the end-session endpoint).
 * Users can now specify the number of days, weeks, and months of site activity to query through the GraphQL API.
+* Added 14 new experimental language servers on the code intelligence admin page.
 
 ## 2.7.6